feat: 调整字段的类型

See merge request ckb-explorer/ckb-explorer-web!36

+++ src/main/java/com/ckb/explorer/entity/StatisticInfo.java
@@ -44,7 +44,7 @@ public class StatisticInfo {
   /**
    * 哈希率（算力）
    */
-  private Double hashRate;
+  private BigDecimal hashRate;
 
   /**
    * 区块链基本信息
